license_portal/api/requirements.txt
Django==2.2.3
psycopg2
djangorestframework==3.10.1
